Create base html, css, js for Etch-a-Sketch
I began with creating the html boiler plate. Then after viewing some examples
of how others set up their html I settled on creating a header section, a
main section, and a footer section. The header section will just display
the name of the app, the main section will contain controls for the grid
and the grid itself, and the footer will have information about copyright
and links.

The css started simple with setting the margin and padding to 0 for the
entire html document. Then I made sure to give my div that will contain the
grid a display of grid. Next I went into js to create the divs within this
container div with a function.

This function takes a gridSize variable, which I have set as 16 for now, and
uses a for loop to create a 16 * 16 size grid of square divs. I achieved this
by using css grid and setting each newly created div's class to square, a
class that gives them a height and width to make them each a grey square.

After tweaking a few things, mainly changing the display from grid to
inline-grid, I finally achieved the etch-a-sketch look with 16 x 16 squares.
Next I added an eventlistener into this function to listen for a mouseenter
event. Upon mouseenter, I added another classlist of 'color' to the div
element. This class in css gives the square div a turquise color for now.

Grid is created and changes color upon mouse scroll.