Add Event Listener/Create Grid
I added an event listener for a click event to the input with SUBMIT value.
This listener will run the userGrid function which displays a text
that shows the user the size of the grid they have selected.

It also runs the createGrid function. This works and changes the size
of the grid. However, it does not erase what was previously
drawn onto the grid. I will have to work on that.