Tweak javascript to add rainbow colorTheme and improve color picker
Created a switch statement has a situation for each value that colorTheme
based on the values attributed to the radio buttons in the html doc.
There is also a eventListener I added to change the value of the colorTheme
based on if there is change in any of the radio buttons.

I also added a for loop in the newGrid() function to uncheck the radio
buttons when the page is reloaded.